brick program (#162)
* brick program

* fmt + clippy

* clippy...again

* should error

---------

+++ token-lending/program/tests/helpers/solend_program_test.rs
@@ -267,7 +267,7 @@ impl SolendProgramTest {
         let instructions = [system_instruction::create_account(
             &self.context.payer.pubkey(),
             &keypair.pubkey(),
-            rent as u64,
+            rent,
             size as u64,
             owner,
         )];
